refac: move task_info from common to witness-gen
