+++ common/src/lib.rs
@@ -1,4 +1,3 @@
 pub mod db;
-pub mod task_info;
 pub mod utils;
 pub mod version;
